test: rename all api to controller
