+++ src/main/kotlin/ecommerce/product/controller/ProductController.kt
@@ -1,29 +1,25 @@
-package ecommerce.product.api
+package ecommerce.product.controller
 
+import ecommerce.product.data.ProductMapper
 import ecommerce.product.data.ProductRequest
 import ecommerce.product.data.ProductResponse
-import ecommerce.product.data.toResponse
 import ecommerce.store.ProductStore
 import org.springframework.dao.DataAccessException
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.PutMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 
 @RestController
-class ProductAPI(private val store: ProductStore) {
+class ProductController(private val store: ProductStore) {
     @PostMapping(PRODUCTS_ENDPOINT)
     fun createProduct(
         @RequestBody request: ProductRequest,
     ): ResponseEntity<ProductResponse> {
         return try {
             val createdProduct = store.create(request)
-            return ResponseEntity.status(HttpStatus.CREATED).body(createdProduct.toResponse())
+            return ResponseEntity
+                .status(HttpStatus.CREATED)
+                .body(ProductMapper.toResponse(createdProduct))
         } catch (exception: DataAccessException) {
             ResponseEntity.status(HttpStatus.BAD_REQUEST).build()
         }
@@ -31,7 +27,13 @@ class ProductAPI(private val store: ProductStore) {
 
     @GetMapping(PRODUCTS_ENDPOINT)
     fun getProducts(): ResponseEntity<List<ProductResponse>> {
-        return ResponseEntity.ok(store.findAll().map { it.toResponse() })
+        return ResponseEntity
+            .ok(
+                store.findAll()
+                    .map {
+                        ProductMapper.toResponse(it)
+                    }
+            )
     }
 
     @PutMapping(PRODUCT_BY_ID_ENDPOINT)
@@ -42,7 +44,7 @@ class ProductAPI(private val store: ProductStore) {
         val updatedProduct = store.update(id, request)
         return when (updatedProduct) {
             null -> ResponseEntity.notFound().build()
-            else -> ResponseEntity.ok(updatedProduct.toResponse())
+            else -> ResponseEntity.ok(ProductMapper.toResponse(updatedProduct))
         }
     }
