+++ src/test/kotlin/ecommerce/product/controller/GetProductControllerTest.kt
@@ -1,4 +1,4 @@
-package ecommerce.product.api
+package ecommerce.product.controller
 
 import ecommerce.product.data.ProductResponse
 import ecommerce.product.helper.TestFixture.ValidRequest.AMERICANO
@@ -15,7 +15,7 @@ import org.springframework.test.annotation.DirtiesContext
 
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-class GetProductAPITest {
+class GetProductControllerTest {
     @Test
     fun `getProducts() should be able to read a product and return 'ok 200' response`() {
         postTestProduct(AMERICANO)
